Move shell types to Microsoft.DotNet.Tools.Bootstrapper.Shell namespace
Move IEnvShellProvider, BashEnvShellProvider, ZshEnvShellProvider,
PowerShellEnvShellProvider, ShellDetection, and ShellProfileManager
into a new Shell/ directory and namespace. These types are used across
multiple commands and don't belong in the PrintEnvScript namespace.
