+++ src/Installer/dotnetup/Shell/ShellProfileManager.cs
@@ -1,9 +1,7 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using Microsoft.DotNet.Tools.Bootstrapper.Commands.PrintEnvScript;
-
-namespace Microsoft.DotNet.Tools.Bootstrapper;
+namespace Microsoft.DotNet.Tools.Bootstrapper.Shell;
 
 /// <summary>
 /// Manages shell profile file modifications to persist .NET environment configuration.
